Delete ocp on cloud parquet too
